FCM 연동 및 Activity 분할
