+++ WarningApp/app/src/main/res/layout/activity_main.xml
@@ -14,31 +14,26 @@
     tools:context=".MainActivity">
 
     <ImageView
+
         android:id="@+id/imageView"
-        android:layout_width="214dp"
-        android:layout_height="230dp"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
         android:background="@color/colorPrimaryDark"
         tools:layout_editor_absoluteX="85dp"
         tools:layout_editor_absoluteY="88dp"
         tools:ignore="MissingConstraints" />
 
     <Button
-        android:id="@+id/selectBtn"
-        android:layout_width="wrap_content"
+        android:id="@+id/signUp"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="이미지선택"
-        tools:layout_editor_absoluteX="148dp"
-        tools:layout_editor_absoluteY="341dp"
-        tools:ignore="MissingConstraints" />
+        android:text="내자녀등록하기" />
 
     <Button
-        android:id="@+id/uploadBtn"
-        android:layout_width="wrap_content"
+        android:id="@+id/violenceList"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="S3업로드"
-        tools:layout_editor_absoluteX="148dp"
-        tools:layout_editor_absoluteY="407dp"
-        tools:ignore="MissingConstraints" />
+        android:text="폭력기록보기" />
 
 
 </LinearLayout>
